fix: remove taskid from scheduled on fetch fail
when a stage is killed because of a fetch fail, we no longer care about
the currently executing tasks of that stage. while we reset their status
we forgot to remove it from the scheduled list. because of which stale
messages were being processed and causing us to enter a invalid state

(L94 scheduler.py) was throwing a KeyError None on task.current